% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bin.r
\name{bin}
\alias{bin}
\title{Create a binned variable.}
\usage{
bin(x, width = find_width(x), origin = find_origin(x, width), name = NULL)
}
\arguments{
\item{x}{numeric or integer vector}

\item{width}{bin width. If not specified, about 10,000 bins will be chosen
using the algorithim in \code{\link{find_width}}.}

\item{origin}{origin. If not specified, guessed by \code{\link{find_origin}}.}

\item{name}{name of original variable. This will be guessed from the input to
\code{group} if not supplied. Used in the output of
\code{\link{condense}} etc.}
}
\description{
Create a binned variable.
}
\details{
This function produces an R reference class that wraps around a C++ function.
Generally, you should just treat this as an opaque object with reference
semantics, and you shouldn't call the methods on it - pass it to
\code{\link{condense}} and friends.
}
\examples{
x <- runif(1e6)
bin(x)
bin(x, 0.01)
bin(x, 0.01, origin = 0.5)
}

