% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ranged.r
\name{is.ranged}
\alias{is.ranged}
\title{Test if an object is of class ranged.}
\usage{
is.ranged(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Test if an object is of class ranged.
}
\keyword{internal}

