\name{norm_deriv}
\alias{norm_deriv}
\title{Compute and sum derivatives of the normal distribution.}
\usage{
  norm_deriv(x, w, order = 1L)
}
\arguments{
  \item{x}{locations at which to evaluate derivatives}

  \item{w}{optional weights}

  \item{order}{order of derivation, currently up to order
  6.}
}
\description{
  Support function for Sheather-Johns plugin estimate of
  the bandwidth.
}
\keyword{internal}

