% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condensed.r
\name{round_any.condensed}
\alias{round_any.condensed}
\title{Round any method for condensed objects}
\usage{
round_any.condensed(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric or date-time (POSIXct) vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\description{
Round any method for condensed objects
}
\keyword{internal}

