\name{scott_breaks}
\alias{scott_breaks}
\title{Pick breaks automatically using Scott's rule}
\usage{
  scott_breaks()
}
\value{
  A function that takes a single parameter, a numeric x
  specifying the data for which breaks are needed, and
  returns a vector of breaks.
}
\description{
  Pick breaks automatically using Scott's rule
}

