\name{smooth_nd_1}
\alias{smooth_nd_1}
\title{Smooth an n-dimensional condensed dataset with normal kernel of
binwidth \code{h} in one direction.}
\usage{
  smooth_nd_1(grid_in, z_in, grid_out, var, h)
}
\description{
  Smooth an n-dimensional condensed dataset with normal
  kernel of binwidth \code{h} in one direction.
}

