\name{condensed}
\alias{condensed}
\title{Condensed data structure.}
\usage{
  condensed(groups, grouped, summary)
}
\arguments{
  \item{groups}{list of \code{\link{grouped}} objects}

  \item{grouped,summary}{output from C++ condense function}
}
\description{
  This object managed the properties of condensed
  (summarised) data frames.
}
\keyword{internal}

