\name{rmse_cvs}
\alias{rmse_cv}
\alias{rmse_cvs}
\title{Estimate smoothing RMSE using leave-one-out cross-valdation.}
\usage{
  rmse_cvs(x, hs = h_grid(x), ...)

  rmse_cv(x, h, var = summary_vars(x)[1], ...)
}
\arguments{
  \item{x}{condensed summary table}

  \item{h,hs}{for \code{rmse_cv}, a vector of bandwidths;
  for \code{rmse_cv} a data frame of bandwidths, as
  generated by \code{\link{h_grid}}.}

  \item{var}{variable to smooth}

  \item{...}{other variables passed on to
  \code{\link{smooth}}}
}
\description{
  Estimate smoothing RMSE using leave-one-out
  cross-valdation.
}
\examples{
set.seed(1014)
# 1d -----------------------------
x <- rchallenge(1e4)
xsum <- condense(bin(x, 1 / 10))
cvs <- rmse_cvs(xsum)

if (require("ggplot2")) {
autoplot(xsum)
qplot(x, err, data = cvs, geom = "line")
xsmu <- smooth(xsum, 1.3)
autoplot(xsmu)
autoplot(peel(xsmu))
}

# 2d -----------------------------
y <- runif(1e4)
xysum <- condense(bin(x, 1 / 10), bin(y, 1 / 100))
cvs <- rmse_cvs(xysum, h_grid(xysum, 10))
if (require("ggplot2")) {
qplot(x, y, data = cvs, size = err)
}
}

